requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.8
django-filter==25.2
django-oauth-toolkit==3.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jwcrypto==1.5.6
oauthlib==3.3.1
packaging==25.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sqlparse==0.5.4
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.0
gunicorn==22.0.0